pixcir: Support inverted IRQ polarity
The TS-TPC-7990 has an inverted interrupt system in order to
allow the touchscreen to wake the device. The CPU pin is expecting
a falling IRQ, but the INT pin on the touchscreen controller must
be set to active high output.

This adds invert-int-output property to the devicetree parsing.
If set, it inverts the INT output from the controller, but the
driver remains configured for active low IRQ.

Base on commit 461fc502098ea5296e0cb3de6afd567a4ce17874 from
https://github.com/embeddedTS/linux-tsimx

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>