clk & pwm: fixups are required for the TS-7680
A combination of commits from the 4.9.y kernel to keep the 12 MHz PWM
clock to the FPGA on after the kernel boots.

pwm-clk is used as a clock source on the TS-7680, however the PWM and clk
drivers all have some shortcomings that have been addressed in this patch.
All of these changes are hidden under a CONFIG_ option that should only
really be enabled for the TS-7680. The PWM is reset, enabled, and set up
in U-Boot.

- Invalid calculation of duty from DTS specified period. At 12 MHz, the
    math the kernel uses breaks down and cannot achieve 12 MHz.
- Various steps of resetting the PWM block or clock gating. All of these
    result in a glitch on the PWM output from the CPU to the FPGA which
    can cause issues if the kernel attempts to talk to the FPGA at this time;
    e.g. if the kernel tries to manipulate a DIO in the FPGA

pwm: Changed TS-7680 fixups

Removed CONFIG option for TS-7680 PWM workaround. This allows defconfig to
be more generic. Instead, the changes in commit 0c03aedeaf66dccbb54 are
triggered based on of_machine_is_compatible specifically for the ts7680
string.