mmc: host: mxs-mmc: handle EPROBE_DEFER from regulator
On platforms with an external GPIO expander that could control
the MMC regulator, need to correctly retry a deferred call.

This changes regulator handling to a mix of what TS did in the
3.14-imx28 kernel and other MMC host driver setup in this repo.
Based on commit 34de4aae1f5e369b0962b2180f6a86506c25e625 from
our linux-4.9.y branch which is based on commit
c0ec175a896b8fbdcea51c8b4318c28057c013fc from 3.14-imx28 and
mmci.c drivers.

Specifically addresses issues with LSR TiWi-BLE module on the
TS-7680.

Implemented what look like current kernel paradigms for probe,
but needed to implement custom regulator handling in set_ios().
This is because ran in to a bit of trouble with SD OCR register
for the voltage bits. Kernel helper functions were returning
EINVAL when attempting to set voltages. In our implementations
the regulators are only ever fixed so this shouldn't be an
issue.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>