serial: 8250: 8250_ts: use mapbase for base addr
This fixes an issue with mapbase/membase/iobase not being used. If none of those are set,
the driver loads and appears to work but never calls uart_config on startup. It then ignores
the value that should be initialized for FCR. This causes the fifo to be disabled.

This driver also now handles the serial in/out hooks in this 8250_ts.c driver.