ARM: dts: ts7400v2: Remove mmc aliases
There is a long history as to why the aliases were there, they
worked on the stock kernel for the TS-7400-V2, did not work in 4.9,
and again work in 5.10.

The issue is that the U-Boot binary expects the eMMC to be at
mmcblk1. This U-Boot was created specifically for the 4.9 kernel.
In 5.10 where the alias was honored, it was showing up at mmcblk2
and resulting in a hang waiting for a non-existant device node.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>