ARM: tsimx6ul: remap fbcon via cmdline extend
While DRM has emulation for fbdev, by default this also enables
FRAMEBUFFER_CONSOLE which cannot be disabled without much additional
work and consideration as it requires CONFIG_EXPERT.

Instead, using fbcon kernel commandline arguments, its possible
to remap the tty to a different /dev/fb* device. In this case, using
fbcon=map:31, we can force a tty to not spawn on any /dev/fb* below
/dev/fb30

In order to work around this across all devices, this cmdline option
is instead moved to the kernel's builtin default cmdline, and the
default cmdline is then set to extend the cmdline passed by the
bootloader.

This is needed as the i.MX6UL platforms span many U-Boot releases
each with their own way of handling the kernel commandline. Adding
the fbcon option to the commandline would require touching multiple
types of U-Boot scripts.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>