net: dsa: mv88e6xxx: add switch-needs-reset OF property
Previously, the probe process expected that the switch:
- Was properly reset and the clock has not turned off since unreset
or
- Is held in a reset state with a valid clock

In cases where the bootloader may not have given a real hardware
reset to the switch, or the clock was turned off between the bootloader
and probe() of this driver, a real reset needs to be given by the
driver before any of the detection calls will work properly.

This real reset is gated behind an OF property so as to not break
existing platforms using this driver that do not want to issue a
hardware reset at the time of early probe.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>