drivers: mmc: tssdcard: fix rare oops on startup
On rare instances, an oops would occur that originated deep within
slab allocation. It appeared to be a race condition that looks like
it was caused by some improper function arguments and calls. This
removes calls to register_blkdev (no longer needed) and passes
correct args to blk_alloc_disk.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>