ARM: dts: imx28-ts7400v2: make i2c0 GPIO
While not an issue with function, using the hardware I2C controller
here would cause a handful of kernel warnings on shutdown due to
the driver not having an atomic handler. Since the WDT via I2C is
involved in halt and reboot, using the peripheral driver would work
but would generate those warnings.

The move to using GPIO pins is mostly aesthetic so as to not have
that warning on halt/reboot.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>