clk & pwm: fixups required for the TS-7680
A combination of commits from 4.9-6.6 kernels to keep the 12 MHz PWM
clock to the FPGA on after the kernel boots.

pwm-clock is used as a clock source on the TS-7680, however the PWM and clk
drivers all have some shortcomings that have been addressed in this patch.
All of these changes are hidden behind matching OF machine compatible
strings to the TS-7680.

The PWM is reset, enabled, and set up in U-Boot; Linux just needs
to let it keep running.

- Invalid calculation of duty from DTS specified period. At 12 MHz, the
    math the kernel uses breaks down and cannot achieve 12 MHz.
- Various steps of resetting the PWM block or clock gating. All of these
    result in a glitch on the PWM output from the CPU to the FPGA which
    can cause issues when later communicating with the FPGA.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>